requirements.txt
torch
torchvision
torchaudio
timm
opencv-python
Pillow
scikit-learn
matplotlib
seaborn
pandas
numpy
mlflow
dvc[s3]
fastapi
uvicorn
python-multipart
boto3
apache-airflow
pytest
albumentations
icrawler
tqdm
flake8
